Added props/targets templates, and also refactored to exporter model
