+++ Source/MSBuildTools.Unity/Packages/com.microsoft.msbuildforunity/Editor/ProjectGenerator/MSBuildTemplates/SDKProjectTemplate.g.props.template.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0140e0bb33cdbb24a91c4201bbdb21e4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
