+++ Source/MSBuildTools.Unity/Packages/com.microsoft.msbuildforunity/Editor/ProjectGenerator/MSBuildTemplates/SDKProjectTemplate.g.targets.template.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bf2e3b7e7e5c0464cb599fda0d57574e
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
