+++ Source/MSBuildTools.Unity/Packages/com.microsoft.msbuildforunity/Editor/ProjectGenerator/Scripts/Exporters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f64b80b10d6120a40972f17a04716012
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
